+++ Modules/Intent.Modules.Eventing.MassTransit/FactoryExtensions/TelemetryConfigurator.cs
@@ -28,40 +28,23 @@ public class TelemetryConfigurator : FactoryExtensionBase
             return;
         }
 
-        UpdateIntegrationEventMessage(application);
         UpdateOpenTelemetryConfiguration(application);
     }
 
     private void UpdateOpenTelemetryConfiguration(IApplication application)
     {
         var template = application.FindTemplateInstance<ICSharpFileBuilderTemplate>(TemplateDependency.OnTemplate("Distribution.OpenTelemetry"));
-        if (template == null)
-        {
-            return;
-        }
 
-        template.CSharpFile.AfterBuild(file =>
+        template?.CSharpFile.AfterBuild(file =>
         {
             var priClass = file.Classes.First();
             var method = priClass.FindMethod("AddTelemetryConfiguration");
-            var telemConfigStmt = (CSharpMethodChainStatement)method.FindStatement(stmt => stmt.HasMetadata("telemetry-config"));
-            var telemTranceStmt = (CSharpInvocationStatement)telemConfigStmt.FindStatement(stmt => stmt.HasMetadata("telemetry-tracing"));
-            var traceChain = (CSharpMethodChainStatement)((CSharpLambdaBlock)telemTranceStmt.Statements.First()).Statements.First();
+            var telemetryConfigStmt = (CSharpMethodChainStatement)method.FindStatement(stmt => stmt.HasMetadata("telemetry-config"));
+            var telemetryTranceStmt = (CSharpInvocationStatement)telemetryConfigStmt.FindStatement(stmt => stmt.HasMetadata("telemetry-tracing"));
+            var traceChain = (CSharpMethodChainStatement)((CSharpLambdaBlock)telemetryTranceStmt.Statements.First()).Statements.First();
 
             traceChain.Statements.Insert(0, @"AddSource(""MassTransit"")");
         });
     }
 
-    private static void UpdateIntegrationEventMessage(IApplication application)
-    {
-        var templates = application.FindTemplateInstances<IntegrationEventMessageTemplate>(TemplateDependency.OnTemplate(IntegrationEventMessageTemplate.TemplateId));
-        foreach (var template in templates)
-        {
-            template.CSharpFile.AfterBuild(file =>
-            {
-                var rec = file.Records.First();
-                rec.Properties.Insert(0, new CSharpProperty(template.UseType("System.Guid") + "?", "CorrelationId", rec).Init());
-            }, 1000);
-        }
-    }
 }